requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
folium
geopandas
joblib
FastAPI
# passlib[bcrypt]
uvicorn
# bcrypt
pymongo
motor 
python-jose[cryptography] 
python-multipart
python-dotenv
pydantic[email]